Update to CXF 3.2.0 snapshot from 2017-07-11
Note that I pulled in additional changes made after
the 20170711 build in the cxf-rt-rs-client module.

These changes are brought in to support the
executorService method tests and to eventually
enable clustering support.

Add failover feature bundle, but don't package it
to avoid missing dependencies.  We'll add those
dependencies in future commits.

Remove references to clustering bundle