Check for references in config when adding features
When a feature is added to the runtime, the metatype definitions in its
bundles are made available, which requires any config elements which
use these definitions to be re-evaluated.

The change in this commit is to also re-evaluate any references to
config elements which use the new definitions.

Previously if you had some config like this:
<myfeature sslRef="mySsl"/>
<ssl id="mySsl"/>

and then enable the ssl-1.0 feature, the sslref on the myfeature PID
would not get updated in config admin.