Issue #787: Fix EJB and Server stopping concurrently
Generally the EJB applications should stop before the server stops,
but if it takes too long there can be ugly errors. There are 3
changes made to avoid the known failures:

1 - If the MBeans have been unregistered already, no sense logging
FFDC for that, just ignore the exceptions

2 - If the EJBRuntime.deactivate method has been called, then
don't bother calling to stop beans, as the internal EJB Container
state has been cleared, and homeOfHomes will be gone

3 - If the EJBRuntime.deactivate has been called, then don't
bother trying to use @Reference fields like the ClassLoaderService,
as those will likely have gone away.

Also, fixed isStopping() to indicate when the server is known to
be stopping, although this doesn't currently appear to be used
in Liberty... and has different meaning than deactivate.