Get configuration from Microprofile Config
Allow annotation parameters and whether annotations are disabled to be
configured via. Microprofile Config.

This replaces the old mechanism which used environment variables and
system properties directly.

Note that we retrieve the config based on the classloader of the
annotated class, not from the thread context classloader as is usual for
Microprofile Config.