Initial app security auto feature
Make JaxRsProviderRegister an interface for OSGi services.
Convert existing SecurityAnnoProviderRegister and
CustomExceptionMapperRegister into declarative services.
Create new AuthenticationFilter to be registered via new
AuthenticationProviderRegister.
AuthenticationFilter will acquire the current security collaborator and
will authenticate the user based on the servlet request/response - this
is only done if the user is not already authenticated.